Refactor forms into separate components and load data from API
